Added New Interactable That Triggers Narration Lines.
Added a new interactable that triggers the next narration line and added some narration lines to test the system.
The interactable as of now has a simple static mesh and text particles for the interaction word.
Also added a spawner for this interactable that spawns a new one  in one of 4 predesignated positions on the map, when the new line plays.